+++ react/components/SKUSelector/Wrapper.tsx
@@ -37,7 +37,13 @@ const getVariationsFromItems = (
   // Transform set back to array
   for (const variationName of variationsNames) {
     const set = variationsSet[variationName]
-    variations[variationName] = Array.from(set)
+    variations[variationName] = {
+      originalName: variationName,
+      values: Array.from(set).map(value => ({
+        name: value,
+        originalName: value,
+      })),
+    }
   }
   return variations
 }
@@ -52,9 +58,13 @@ const getVariationsFromSpecifications = (
       !visibleVariations ||
       visibleVariations.includes(specification.field.name.toLowerCase().trim())
     ) {
-      variations[specification.field.name] = specification.values.map(
-        value => value.name
-      )
+      variations[specification.field.name] = {
+        originalName: specification.field.originalName,
+        values: specification.values.map(value => ({
+          name: value.name,
+          originalName: value.originalName,
+        })),
+      }
     }
   }
   return variations
@@ -72,6 +82,7 @@ const useVariations = (
   const variationsSource = isSkuSpecificationsEmpty
     ? skuItems
     : skuSpecifications
+
   const result = useMemo(() => {
     if (
       shouldNotShow ||
@@ -101,6 +112,7 @@ const useVariations = (
     visibleVariations,
     isSkuSpecificationsEmpty,
   ])
+
   return result
 }